backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.13.1
amqp==5.2.0
aniso8601==9.0.1
anyio==4.2.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.14.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
beniget==0.4.1
billiard==4.2.0
bleach==6.1.0
blinker==1.7.0
Brlapi==0.8.3
Brotli==1.0.9
cachelib==0.9.0
cachetools==5.0.0
calplot==0.1.7.5
celery==5.4.0
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
checksumdir==1.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
comm==0.2.0
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
duplicity==0.8.21
email_validator==2.1.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fasteners==0.14.1
fastjsonschema==2.19.1
filelock==3.15.1
Flask==3.0.0
Flask-Caching==2.3.0
Flask-Cors==4.0.1
Flask-Excel==0.0.7
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.29.1
fqdn==1.5.1
frozenlist==1.2.0
fs==2.4.12
fsspec==2024.10.0
future==0.18.2
gast==0.5.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
greenlet==3.0.3
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
indicator-sound-switcher==2.3.10.1
ipykernel==6.28.0
ipython==8.19.0
ipython_genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.1
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupyterthemes==0.20.0
keyring==23.5.0
kiwisolver==1.3.2
kombu==5.3.7
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lesscpy==0.15.1
lml==0.1.0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.3
matplotlib==3.5.1
matplotlib-inline==0.1.6
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.3
multidict==5.1.0
nbclient==0.9.0
nbconvert==7.13.1
nbformat==5.9.2
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.4.2
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.5
overrides==7.4.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==4.1.0
ply==3.11
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==3.12.4
proton-core==0.3.3
proton-keyring-linux==0.1.0
proton-keyring-linux-secretservice==0.1.0
proton-vpn-api-core==0.36.4
proton-vpn-gtk-app==4.7.1
proton-vpn-killswitch-network-manager==0.6.1
proton-vpn-killswitch-network-manager-wireguard==0.2.1
proton-vpn-logger==0.2.1
proton-vpn-network-manager==0.9.4
proton-vpn-network-manager-openvpn==0.1.1
proton-vpn-network-manager-wireguard==0.4.7
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
PyDrive==1.3.1
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-json-logger==2.0.7
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
redis==5.0.6
referencing==0.32.0
reportlab==3.6.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.16.2
rsa==4.8
scikit-learn==1.4.0
scipy==1.8.0
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.2
sentry-sdk==1.4.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
ssh-import-id==5.11
stack-data==0.6.3
sympy==1.13.1
systemd-python==234
tabulate==0.8.9
terminado==0.18.0
testpath==0.5.0
texttable==1.7.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tornado==6.4
traitlets==5.14.0
triton==3.1.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.4
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uri-template==1.3.0
uritemplate==3.0.1
urllib3==1.26.5
usb-creator==0.3.7
vine==5.1.0
virtualenv==20.26.2
wadllib==1.3.6
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
WTForms==3.1.2
xdg==5
xkit==0.0.0
yarl==1.7.2
zipp==1.0.0
